Restructure equations. Isolate correlation-specific concerns from implicit equations data schema
