--- changelog/10.change.md
@@ -1,1 +1,0 @@
-Structure equations for testability
