+++ changelog/182.change.md
@@ -1,0 +1,1 @@
+Structure equations for testability. Add dimensionless parameters to `boilercv`.
