+++ data/dimensionless_bubble_diameter_equations.toml
@@ -19,14 +19,14 @@ sym = 'Pr'
 test = 1.0
 
 [[equation]]
-name = 'dimensionless_bubble_diameter_florschuetz_chao_1965'
+name = 'florschuetz_chao_1965'
 latex = '\beta = 1 - 4 \Ja \sqrt{\Fo_\o / \pi}'
 sympy = 'Eq(beta, 1 - 4 * Ja * sqrt(Fo_0 / pi))'
 python = '1 - 4 * bubble_jakob * sqrt(bubble_fourier / pi)'
 expect = [1.000000, 0.946807, 0.924774, 0.907868, 0.893615, 0.881058, 0.869705, 0.859266, 0.849549, 0.840423]
 
 [[equation]]
-name = 'dimensionless_bubble_diameter_isenberg_sideman_1970'
+name = 'isenberg_sideman_1970'
 latex = '''
 \beta =
 \left(1 - \left(3/\sqrt{\pi}\right) \Re_\bo^{1/2} \Pr^{1/3} Ja Fo_\o \right)^{2/3}
@@ -41,7 +41,7 @@ python = '''
 expect = [1.000000, 0.993721, 0.987422, 0.981104, 0.974765, 0.968405, 0.962024, 0.955622, 0.949199, 0.942753]
 
 [[equation]]
-name = 'dimensionless_bubble_diameter_akiyama_1973'
+name = 'akiyama_1973'
 latex = '\beta=(1 - 1.036 Re_\bo^{1/2} Pr^{1/3} Ja Fo_\o)^{0.714}'
 sympy = 'Eq(beta, (1 - 1.036 * Re_b0**(1/2) * Pr**(1/3) * Ja * Fo_0)**0.714)'
 python = '''
@@ -53,7 +53,7 @@ python = '''
 expect = [1.000000, 0.995887, 0.991767, 0.987640, 0.983507, 0.979367, 0.975219, 0.971065, 0.966903, 0.962734]
 
 [[equation]]
-name = 'dimensionless_bubble_diameter_yuan_et_al_2009'
+name = 'yuan_et_al_2009'
 latex = '''
 \beta =
 \left(
