--- pipeline/boilercv_pipeline/correlations/symbolic/__init__.py
@@ -1,5 +1,0 @@
-"""Symbolic correlations."""
-
-from boilercv_pipeline.correlations.symbolic import (  # noqa: F401
-    dimensionless_bubble_diameter,
-)
