+++ tests/boilercv_tests/test_equations/__init__.py
@@ -1,0 +1,1 @@
+"""Equations tests."""
